fixes for home_19 and home_20
