+++ home_19/Settings/TestRunParameters.cs
@@ -1,0 +1,12 @@
+﻿using System.Xml.Serialization;
+
+namespace home_19.Settings
+{
+    [XmlRoot(ElementName = "TestRunParameters")]
+    public class TestRunParameters
+    {
+
+        [XmlElement(ElementName = "Parameter")]
+        public List<Parameter> Parameter { get; set; }
+    }
+}
